+++ codex-rs/core/src/config/config_tests.rs
@@ -5,9 +5,6 @@ use assert_matches::assert_matches;
 use codex_config::CONFIG_TOML_FILE;
 use codex_config::ConfigLayerEntry;
 use codex_config::ConfigLayerStack;
-use codex_config::ConfigRequirements;
-use codex_config::ConfigRequirementsToml;
-use codex_config::FeatureRequirementsToml;
 use codex_config::ProfileV2Name;
 use codex_config::RequirementSource;
 use codex_config::Sourced;
@@ -38,7 +35,6 @@ use codex_config::permissions_toml::PermissionsToml;
 use codex_config::permissions_toml::WorkspaceRootsToml;
 use codex_config::types::AppToolApproval;
 use codex_config::types::ApprovalsReviewer;
-use codex_config::types::AuthKeyringBackendKind;
 use codex_config::types::BundledSkillsConfig;
 use codex_config::types::FeedbackConfigToml;
 use codex_config::types::HistoryPersistence;
@@ -5060,74 +5056,6 @@ async fn feature_table_overrides_legacy_flags() -> std::io::Result<()> {
     assert!(!config.features.enabled(Feature::ApplyPatchFreeform));
 
     Ok(())
-}
-
-#[test]
-fn resolve_bootstrap_auth_keyring_backend_kind_uses_secret_auth_storage_feature()
--> std::io::Result<()> {
-    let config_toml = ConfigToml {
-        features: Some(FeaturesToml::from(BTreeMap::from([(
-            "secret_auth_storage".to_string(),
-            true,
-        )]))),
-        ..Default::default()
-    };
-    assert_eq!(
-        resolve_bootstrap_auth_keyring_backend_kind(&config_toml_load_result(
-            config_toml,
-            /*feature_requirements*/ None,
-        )?)?,
-        AuthKeyringBackendKind::Secrets
-    );
-
-    let config_toml = ConfigToml {
-        features: Some(FeaturesToml::from(BTreeMap::from([(
-            "secret_auth_storage".to_string(),
-            false,
-        )]))),
-        ..Default::default()
-    };
-    assert_eq!(
-        resolve_bootstrap_auth_keyring_backend_kind(&config_toml_load_result(
-            config_toml.clone(),
-            /*feature_requirements*/ None,
-        )?)?,
-        AuthKeyringBackendKind::Direct
-    );
-
-    let requirements = Sourced::new(
-        FeatureRequirementsToml {
-            entries: BTreeMap::from([("secret_auth_storage".to_string(), true)]),
-        },
-        RequirementSource::Unknown,
-    );
-    assert_eq!(
-        resolve_bootstrap_auth_keyring_backend_kind(&config_toml_load_result(
-            config_toml,
-            Some(requirements),
-        )?)?,
-        AuthKeyringBackendKind::Secrets
-    );
-
-    Ok(())
-}
-
-fn config_toml_load_result(
-    config_toml: ConfigToml,
-    feature_requirements: Option<Sourced<FeatureRequirementsToml>>,
-) -> std::io::Result<ConfigTomlLoadResult> {
-    let requirements = ConfigRequirements {
-        feature_requirements,
-        ..Default::default()
-    };
-    Ok(ConfigTomlLoadResult {
-        config_toml,
-        config_layer_stack: ConfigLayerStack::new(
-            Vec::new(),
-            requirements,
-            ConfigRequirementsToml::default(),
-        )?,
-    })
 }
 
 #[tokio::test]
